+++ foundation-service/src/main/java/io/soffa/foundation/service/core/config/SecurityConfig.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -15,9 +16,6 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
 
 @Configuration
 @EnableWebSecurity
@@ -36,19 +34,6 @@ public class SecurityConfig {
     }
 
     @Bean
-    @ConditionalOnMissingBean(CorsFilter.class)
-    public CorsFilter corsFilter() {
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        CorsConfiguration config = new CorsConfiguration();
-        config.setAllowCredentials(true);
-        config.addAllowedOriginPattern("*");
-        config.addAllowedHeader("*");
-        config.addAllowedMethod("*");
-        source.registerCorsConfiguration("/**", config);
-        return new CorsFilter(source);
-    }
-
-    @Bean
     @ConditionalOnMissingBean
     public UserDetailsService userDetailsService() {
         return new InMemoryUserDetailsManager();
@@ -56,28 +41,27 @@ public class SecurityConfig {
 
     @Bean
     public SecurityFilterChain configure(HttpSecurity http) throws Exception {
+
         http
-            .cors(AbstractHttpConfigurer::disable)
+            .cors(Customizer.withDefaults())
             .csrf(AbstractHttpConfigurer::disable)
             .httpBasic(AbstractHttpConfigurer::disable)
             .formLogin(AbstractHttpConfigurer::disable)
-            .sessionManagement((session) -> session
+            .sessionManagement(session -> session
                 .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
             )
-            .authorizeHttpRequests((requests) -> {
-                if ("authenticated" .equalsIgnoreCase(openApiAccess)) {
-                    requests
-                        .requestMatchers("/v3/**", "/swagger/*").authenticated();
+            .authorizeHttpRequests(requests -> {
+                if ("authenticated".equalsIgnoreCase(openApiAccess)) {
+                    requests.requestMatchers("/v3/**", "/swagger/*").authenticated();
                 } else {
-                    requests
-                        .requestMatchers("/v3/**", "/swagger/*").permitAll();
+                    requests.requestMatchers("/v3/**", "/swagger/*").permitAll();
                 }
 
-                requests.requestMatchers("/actuator/**").permitAll()
+                requests
+                    .requestMatchers("/actuator/**").permitAll()
                     .requestMatchers("/**").permitAll()
                     .anyRequest()
                     .authenticated();
-
             })
             .addFilterBefore(new RequestFilter(authManager), UsernamePasswordAuthenticationFilter.class);